+++ lib/rubocop/cop/lint/number_conversion.rb
@@ -14,12 +14,14 @@ module RuboCop
       #   '10'.to_i
       #   '10.2'.to_f
       #   '10'.to_c
+      #   ['1', 2, 3, 4].map(&:to_i)
       #
       #   # good
       #
       #   Integer('10', 10)
       #   Float('10.2')
       #   Complex('10')
+      #   ['1', 2, 3, 4].map { |i| Integer(i, 10) }
       class NumberConversion < Cop
         CONVERSION_METHOD_CLASS_MAPPING = {
           to_i: "#{Integer.name}(%<number_object>s, 10)",
@@ -28,11 +30,15 @@ module RuboCop
         }.freeze
         MSG = 'Replace unsafe number conversion with number '\
               'class parsing, instead of using '\
-              '%<number_object>s.%<to_method>s, use stricter '\
-              '%<corrected_method>s.'
+              '%<violating_expression>s, use stricter '\
+              '%<corrected_expression>s.'
 
         def_node_matcher :to_method, <<~PATTERN
-          (send $_ ${:to_i :to_f :to_c})
+          (send $_ {:to_i :to_f :to_c})
+        PATTERN
+
+        def_node_matcher :to_method_proc?, <<~PATTERN
+          (send $_ ${:map :map! :collect :collect!} (block-pass (sym ${:to_i :to_f :to_c})))
         PATTERN
 
         def_node_matcher :datetime?, <<~PATTERN
@@ -40,27 +46,49 @@ module RuboCop
         PATTERN
 
         def on_send(node)
-          to_method(node) do |receiver, to_method|
-            next if receiver.nil? || date_time_object?(receiver)
-
-            message = format(
-              MSG,
-              number_object: receiver.source,
-              to_method: to_method,
-              corrected_method: correct_method(node, receiver)
-            )
-            add_offense(node, message: message)
-          end
+          check_to_method(node)
+          check_to_method_proc(node)
         end
 
         def autocorrect(node)
           lambda do |corrector|
-            corrector.replace(node,
-                              correct_method(node, node.receiver))
+            if to_method_proc?(node)
+              corrector.replace(node, corrected_block_expression(*to_method_proc?(node)))
+            else
+              corrector.replace(node, to_method_corrected(node, node.receiver))
+            end
           end
         end
 
         private
+
+        def check_to_method(node)
+          receiver = to_method(node)
+
+          return unless receiver
+
+          return if receiver.nil? || date_time_object?(receiver)
+
+          message = format(
+            MSG,
+            violating_expression: node.source,
+            corrected_expression: to_method_corrected(node, receiver)
+          )
+          add_offense(node, message: message)
+        end
+
+        def check_to_method_proc(node)
+          captured_values = to_method_proc?(node)
+
+          return unless captured_values
+
+          msg = format(
+            MSG,
+            violating_expression: node.source,
+            corrected_expression: corrected_block_expression(*captured_values)
+          )
+          add_offense(node, message: msg)
+        end
 
         def date_time_object?(node)
           child = node
@@ -71,7 +99,21 @@ module RuboCop
           end
         end
 
-        def correct_method(node, receiver)
+        def corrected_block_expression(receiver, operation, to_method)
+          corrected_block = format(
+            "{ |%<number_object>s| #{CONVERSION_METHOD_CLASS_MAPPING[to_method]} }",
+            number_object: 'i'
+          )
+
+          format(
+            '%<receiver>s.%<operation>s %<corrected_block>s',
+            receiver: receiver.source,
+            operation: operation,
+            corrected_block: corrected_block
+          )
+        end
+
+        def to_method_corrected(node, receiver)
           format(CONVERSION_METHOD_CLASS_MAPPING[node.method_name],
                  number_object: receiver.source)
         end