+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [#8017](https://github.com/rubocop-hq/rubocop/pull/8017): Fix a false positive for `Lint/SuppressedException` when empty rescue with comment in `def`. ([@koic][])
 * [#7990](https://github.com/rubocop-hq/rubocop/issues/7990): Fix resolving `inherit_gem` in remote configs. ([@CvX][])
 * [#8035](https://github.com/rubocop-hq/rubocop/issues/8035): Fix a false positive for `Lint/DeprecatedOpenSSLConstant` when using double quoted string argument. ([@koic][])
+* [#7902](https://github.com/rubocop-hq/rubocop/issues/7902): Handle unsafe conversion methods as symbols for `map` and `collect` in Lint/NumberConversion. ([@petar-lazarov][])
 
 ## 0.84.0 (2020-05-21)
 
@@ -4542,3 +4543,4 @@
 [@laurmurclar]: https://github.com/laurmurclar
 [@jethrodaniel]: https://github.com/jethrodaniel
 [@CvX]: https://github.com/CvX
+[@petar-lazarov]: https://github.com/petar-lazarov